+++ Deposits.go
@@ -19,7 +19,7 @@ const TYPE_DEPOSIT = "Deposit"
 
 type Deposits struct {
     Id                   uint    `gorm:"primary_key"`
-    Currency             uint8
+    Currency             uint
     Amount               decimal.NullDecimal `sql:"type:decimal(32,16);"`
     Fee                  decimal.NullDecimal `sql:"type:decimal(32,16);"`
     FundUid              string  `gorm:"size:255" sql:"default: null"`
